requirements.txt
coreapi==2.1.1
Django==1.10.3
django-cors-headers==1.3.1
django-rest-swagger==2.1.0
djangorestframework==3.5.3
djangorestframework-jsonapi==2.1.1
djangorestframework-jwt==1.9.0
inflection==0.3.1
itypes==1.1.0
openapi-codec==1.2.1
PyJWT==1.4.2
python-ldap==2.4.28
requests==2.12.3
simplejson==3.10.0
uritemplate==3.0.0
